% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file_list.R
\name{check_file_list}
\alias{check_file_list}
\title{check_file_list: Verifica se uma lista de nomes de arquivos está completa}
\usage{
check_file_list(
  dir_files,
  data_type,
  state_abbr,
  publication_date_start,
  publication_date_end
)
}
\arguments{
\item{dir_files}{Um DataFrame contendo os nomes dos arquivos e a data correspondente a cada arquivo}

\item{data_type}{String. Valores aceitos "PA", "RD", "RJ", "SP"}

\item{state_abbr}{String. Sigla da Unidade Federativa}

\item{publication_date_start}{Uma string no formato "AAAA-MM-01", indicando o mes de inicio para o download dos dados.}

\item{publication_date_end}{Uma string no formato "AAAA-MM-01", indicando o mes de termino para o download dos dados.}
}
\description{
A função check_file_list verifica se a função list_SIA_SIH_files retornou um arquivo para cada mês. Se algum mês estiver sem arquivo, a função imprime no console as datas dos meses faltantes.
}
\examples{
\dontrun{
  check_file_list(
    dir_files = dir_files,
    data_type = "PA",
    state_abbr = "CE",
    publication_date_start = "2023-01-01",
    publication_date_end = "2023-03-01"
  )
}

}
