% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_3_check_filter.R
\name{check_filter}
\alias{check_filter}
\title{Receives a database and returns a filtered version of it.}
\usage{
check_filter(df, var_value, var_name)
}
\arguments{
\item{df}{um DataFrame que sera filtrado.}

\item{var_value}{string ou vetor de strings. Os valores a serem filtrados. Se o valor for nulo, a função retorna o mesmo dataframe recebido.}

\item{var_name}{string. Nome da coluna do DataFrame que será utilizada para a filtragem.}
}
\value{
Retorna uma tabela filtrada pelo valor de \code{var_value} na coluna \code{var_name}.
}
\description{
Função auxiliar que filtra uma base de dados com base em um valor específico.
}
