% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_create_output.R
\name{create_output}
\alias{create_output}
\title{Create SUS-SIA/SIH database}
\usage{
create_output(
  year_start,
  month_start,
  year_end,
  month_end,
  state_abbr,
  county_id = "all",
  health_establishment_id = "all"
)
}
\arguments{
\item{year_start}{Ano de início da realização do procedimento}

\item{month_start}{Mês de início da realização do procedimento}

\item{year_end}{Ano de término da realização do procedimento}

\item{month_end}{Mês de término da realização do procedimento}

\item{state_abbr}{Sigla da Unidade Federativa}

\item{county_id}{Codigo(s) do Municipio de Atendimento}

\item{health_establishment_id}{Código(s) do estabelecimento de saúde}
}
\description{
\link{create_output} Processar arquivos dos sistemas de informação SIA e SIH (DATASUS) e combiná-los com informações do CNES e SIGTAP.
}
\examples{
\dontrun{ dados = create_output(year_start=2023,
                  month_start=1,
                  year_end=2023,
                  month_end=3,
                  state_abbr="CE",
                  county_id="230440",
                  health_establishment_id=c("2561492","2481286")
                  )}
}
