% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_2_create_output_from_local.R
\name{create_output_from_local}
\alias{create_output_from_local}
\title{Create SUS-SIA/SIH database from local DBC files}
\usage{
create_output_from_local(
  state_abbr,
  dbc_dir_path,
  county_id = "all",
  health_establishment_id = "all"
)
}
\arguments{
\item{state_abbr}{String. Sigla da Unidade Federativa}

\item{dbc_dir_path}{Diretório que contêm os arquivos DBC}

\item{county_id}{Codigo(s) do Municipio de Atendimento. O padrao é 'all'.}

\item{health_establishment_id}{Código(s) do estabelecimento de saúde}
}
\description{
Processar arquivos dos sistemas de informação SIA e SIH (DATASUS) que já estão baixados localmente e combiná-los com informações do CNES e SIGTAP.
}
\examples{
  \dontrun{
    create_output_from_local(
      state_abbr = "CE",
      dbc_dir_path = "X:/USID/BOLSA_EXTENSAO_2024/dbc/dbc-2301-2306",
      county_id = "230440",
      health_establishment_id = c("2561492", "2481286")
    )
  }

}
