% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_1_create_output_SIH_RJ.R
\name{create_output_SIH_RJ}
\alias{create_output_SIH_RJ}
\title{Create a database for the Hospital Information System (SIH/SUS) - AIH-RJ}
\usage{
create_output_SIH_RJ(
  year_start,
  month_start,
  year_end,
  month_end,
  state_abbr,
  county_id = NULL,
  health_establishment_id = NULL
)
}
\arguments{
\item{year_start}{Um numero de 4 digitos, indicando o ano de inicio para o download dos dados.}

\item{month_start}{Um numero de 2 digitos, indicando o mes de inicio para o download dos dados.}

\item{year_end}{Um numero de 4 digitos, indicando o ano de termino para o download dos dados.}

\item{month_end}{Um numero de 2 digitos, indicando o mes de termino para o download dos dados.}

\item{state_abbr}{String. Sigla da Unidade Federativa}

\item{county_id}{Codigo do Municipio de Atendimento. O padrao é NULL. É obrigatório se health_establishment_id for NULL.}

\item{health_establishment_id}{Código(s) do estabelecimento de saúde. O padrao é NULL. É obrigatório se county_id for NULL}
}
\value{
Um DataFrame estruturado contendo dados do SUS-SIH-AIH-RJ, filtrado por estado ou estabelecimentos de saúde dentro de um intervalo de datas específico, e combinado com informações do CNES e SIGTAP.
}
\description{
Processar arquivos de Autorização de Internação Hospitalar (AIH) Rejeitadas (Rj) do Sistema de Informação Hospitalar (SIH) do DATASUS e integrá-los com dados do CNES e SIGTAP.
}
\examples{
  dados = create_output_SIH_RJ(
    year_start = 2023,
    month_start = 1,
    year_end = 2023,
    month_end = 3,
    state_abbr = "CE",
    county_id = "230440",
    health_establishment_id = c("2561492", "2481286")
  )

}
