% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_2_create_output_SIH_RJ_from_local.R
\name{create_output_SIH_RJ_from_local}
\alias{create_output_SIH_RJ_from_local}
\title{Create a database for the Hospital Information System (SIH/SUS) - AIH-RJ, local data.}
\usage{
create_output_SIH_RJ_from_local(
  state_abbr,
  dbc_dir_path,
  county_id = NULL,
  health_establishment_id = NULL
)
}
\arguments{
\item{state_abbr}{string or a vector of strings. Sigla da Unidade Federativa}

\item{dbc_dir_path}{string. Caminho para o diretório  local que contêm os arquivos DBC}

\item{county_id}{string or a vector of strings. Código do Município de Atendimento. O padrão é NULL. É obrigatório se health_establishment_id for NULL.}

\item{health_establishment_id}{string or a vector of strings. Código do estabelecimento de saúde. O padrao é NULL. É obrigatório se county_id for NULL}
}
\value{
Um DataFrame estruturado contendo dados do SUS-SIH-AIH-RJ, filtrado por estado ou estabelecimentos de saúde dentro de um intervalo de datas específico, e combinado com informações do CNES e SIGTAP.
}
\description{
Processar arquivos de Autorização de Internação Hospitalar (AIH) Rejeitadas (RJ) do Sistema de Informação Hospitalar (SIH) do DATASUS que já estão baixados localmente e integrá-los com dados do CNES e SIGTAP.
}
\examples{
  \dontrun{
    data = create_output_SIH_RJ_from_local(
      state_abbr = "CE",
      dbc_dir_path = "X:/USID/BOLSA_EXTENSAO_2024/dbc/dbc-2301-2306",
      county_id = "230440",
      health_establishment_id = c("2561492", "2481286")
    )
  }

}
