% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_2_create_output_SP_from_local.R
\name{create_output_SP_from_local}
\alias{create_output_SP_from_local}
\title{Create a database for the Hospital Information System (SIH/SUS) - SP, local data.}
\usage{
create_output_SP_from_local(
  state_abbr,
  dbc_dir_path,
  county_id = NULL,
  health_establishment_id = NULL
)
}
\arguments{
\item{state_abbr}{String. Sigla da Unidade Federativa}

\item{dbc_dir_path}{Diretório que contêm os arquivos DBC}

\item{county_id}{Codigo do Municipio de Atendimento. O padrao é NULL. É obrigatório se health_establishment_id for NULL.}

\item{health_establishment_id}{Código(s) do estabelecimento de saúde. O padrao é NULL. É obrigatório se county_id for NULL}
}
\value{
Um DataFrame estruturado contendo dados do SUS-SIH-SP, filtrado por estado ou estabelecimentos de saúde dentro de um intervalo de datas específico, e combinado com informações do CNES e SIGTAP.
}
\description{
Processar arquivos do Sistema de Informação Hospitalar (SIH) dos Serviço Profissionais (SP) do DATASUS que já estão baixados localmente e integrá-los com dados do CNES e SIGTAP.
}
\examples{
    dados = create_output_SP_from_local(
      state_abbr = "CE",
      dbc_dir_path = "X:/USID/BOLSA_EXTENSAO_2024/dbc/dbc-2301-2306",
      county_id = "230440",
      health_establishment_id = c("2561492", "2481286")
    )

}
