% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_1_download_cnes_files.R
\name{download_cnes_files}
\alias{download_cnes_files}
\title{download_cnes_files: Faz o download dos dados do CNES}
\usage{
download_cnes_files(
  year_start,
  month_start,
  year_end,
  month_end,
  newer,
  state_abbr = "all"
)
}
\arguments{
\item{year_start}{Um numero de 4 digitos, indicando o ano de inicio para o download dos dados.}

\item{month_start}{Um numero de 2 digitos, indicando o mes de inicio para o download dos dados.}

\item{year_end}{Um numero de 4 digitos, indicando o ano de termino para o download dos dados.}

\item{month_end}{Um numero de 2 digitos, indicando o mes de termino para o download dos dados.}

\item{newer}{Logico. O padrao e TRUE. Se for TRUE e os outros parametros forem nulos, obtem o arquivo do mes mais recente disponivel no SIGTAP.}

\item{state_abbr}{String. O padrao e "all". Sigla da Unidade Federativa}
}
\description{
A funcao \code{download_cnes_files} baixa arquivos do CNES (Cadastro Nacional de Estabelecimentos de Saude) e os salva temporariamente na maquina local. Esses arquivos podem ser filtrados pela unidade da federacao.
}
\examples{
\dontrun{download_cnes_files(newer = TRUE)}
}
