% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_download_sigtap_files.R
\name{download_sigtap_files}
\alias{download_sigtap_files}
\title{Download SIGTAP Files}
\usage{
download_sigtap_files(
  year_start,
  month_start,
  year_end,
  month_end,
  newer = TRUE,
  specific_dates = NULL
)
}
\arguments{
\item{year_start}{Um numero de 4 digitos, indicando o ano de inicio para o download dos dados.}

\item{month_start}{Um numero de 2 digitos, indicando o mes de inicio para o download dos dados.}

\item{year_end}{Um numero de 4 digitos, indicando o ano de termino para o download dos dados.}

\item{month_end}{Um numero de 2 digitos, indicando o mes de termino para o download dos dados.}

\item{newer}{Logico. O padrao é TRUE. Se for TRUE e os outros parametros forem nulos, obtem o arquivo do mes mais recente disponivel no SIGTAP.}

\item{specific_dates}{Um vetor de strings contendo as datas específicas que se deseja baixar.}
}
\description{
A função \code{download_sigtap_files} baixa e descompacta arquivos SIGTAP do site DataSUS. Esses arquivos são salvos temporariamente na máquina local. Para cada mês, será baixado um arquivo compactado. Primeiro, a função verifica se o arquivo do mês especificado já existe na pasta temporária. Se o arquivo estiver presente, ela não baixa o arquivo novamente para esse mês.
}
\examples{
\dontrun{
   download_sigtap_files(2024, 1, 2024, 3, newer = FALSE)
 }
\dontrun{
   download_sigtap_files(specific_dates = c("202210","202301","202303"))}

}
