% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_3_get_CNES.R
\name{get_CNES}
\alias{get_CNES}
\title{Loads data from the National Registry of Health Establishments (CNES)}
\usage{
get_CNES()
}
\description{
A função cria e salva no ambiente global do R o DataFrame \code{health_establishment}, que contém a união de todos os arquivos baixados pela função \code{download_cnes_files.}
}
