% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_2_get_counties.R
\name{get_counties}
\alias{get_counties}
\title{get_counties: Obtem informacoes de todos os estados do Brasil}
\usage{
get_counties(state_abbr, county_id)
}
\arguments{
\item{state_abbr}{Sigla da Unidade Federativa}

\item{county_id}{Codigo(s) do Municipio de Atendimento}
}
\value{
A funcao nao retorna valores diretamente, mas salva no ambiente global do R os DataFrames \code{counties} e \code{health_establishment}.
}
\description{
A funcao \code{get_counties} cria dois DataFrames. O primeiro, \code{counties}, contem informacoes detalhadas sobre os municipios do Brasil. O segundo, \code{health_establishment}, e a uniao de todos os arquivos baixados pela funcao \code{download_cnes_files}.
}
