% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_get_datasus.R
\name{get_datasus}
\alias{get_datasus}
\title{get_datasus: Chamar e executar todas as outras funções definidas no pacote.}
\usage{
get_datasus(
  year_start,
  month_start,
  year_end,
  month_end,
  state_abbr,
  county_id,
  information_system,
  health_establishment_id
)
}
\arguments{
\item{year_start}{Um numero de 4 digitos, indicando o ano de inicio para o download dos dados.}

\item{month_start}{Um numero de 2 digitos, indicando o mes de inicio para o download dos dados.}

\item{year_end}{Um numero de 4 digitos, indicando o ano de termino para o download dos dados.}

\item{month_end}{Um numero de 2 digitos, indicando o mes de termino para o download dos dados.}

\item{state_abbr}{String. Sigla da Unidade Federativa}

\item{county_id}{Codigo(s) do Municipio de Atendimento}

\item{information_system}{String. Valores aceitos "SIA", "SIH-AIH", "SIH-SP"}

\item{health_establishment_id}{COdigo(s) do estabelecimento de saude}
}
\value{
output
}
\description{
get_datasus: Chamar e executar todas as outras funções definidas no pacote.
}
