% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/6_2_get_datasus_from_local.R
\name{get_datasus_from_local}
\alias{get_datasus_from_local}
\title{get_datasus_from_local: Chamar e executar todas as outras funções definidas no pacote.}
\usage{
get_datasus_from_local(
  state_abbr,
  county_id,
  dbc_dir_path,
  information_system,
  health_establishment_id
)
}
\arguments{
\item{state_abbr}{String. Sigla da Unidade Federativa}

\item{county_id}{Codigo(s) do Municipio de Atendimento}

\item{dbc_dir_path}{Caminho para o diretorio dos arquivos .dbc}

\item{information_system}{String. Valores aceitos "SIA", "SIH-AIH", "SIH-SP"}

\item{health_establishment_id}{Codigo(s) do estabelecimento de saude}
}
\value{
output
}
\description{
get_datasus_from_local: Chamar e executar todas as outras funções definidas no pacote.
}
