% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_2_get_details.R
\name{get_detail}
\alias{get_detail}
\title{Returns a table with SIGTAP data filtered and organized.}
\usage{
get_detail(detail_name)
}
\arguments{
\item{detail_name}{String. Nome do arquivo desejado. Existem 8 opcoes possiveis: "Procedimento", "Grupo", "Subgrupo", "Forma de organizacao", "Financiamento", "Rubrica", "CBO", "CID".}
}
\value{
Uma tabela dos dados correpondentes ao parametro \code{detail_name}
}
\description{
Esta função filtra os dados baixados pela função \code{download_sigtap_files} usando o valor especificado no parâmetro \code{detail_name} e retorna os dados filtrados em um formato estruturado e organizado.
}
