% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_1_get_procedure_details.R
\name{get_procedure_details}
\alias{get_procedure_details}
\title{get_procedure_details: Estrutura os dados baixados pelo \code{download_sigtap_files}}
\usage{
get_procedure_details()
}
\value{
Uma unica tabela contendo alguns dados baixados do SIGTAP
}
\description{
Esta funcao agrupa alguns arquivos baixados previamente pela funcao \code{download_sigtap_files}, consolidando todos os dados em uma unica tabela.
}
\examples{
\dontrun{get_procedure_details()}
}
