% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_2_preprocess_SIH.R
\name{preprocess_SIH}
\alias{preprocess_SIH}
\title{preprocess_SIH: Estrutura os dados de Producao Hospitalar (SIH)}
\usage{
preprocess_SIH(
  cbo,
  cid,
  raw_SIH,
  file_type,
  county_id,
  procedure_details,
  health_establishment_id
)
}
\arguments{
\item{cbo}{E a tabela retornada pela funcao \code{get_details} com o parametro detail_name igual a "CBO"}

\item{cid}{E a tabela retornada pela funcao \code{get_details} com o parametro detail_name igual a "CID".}

\item{raw_SIH}{Dados do DataSUS de Producao Hospitalar (SIH)}

\item{file_type}{tibble, nomes de arquivos .dbc, e algumas descricoes desses arquivos}

\item{county_id}{Codigo(s) do Municipio de Atendimento}

\item{procedure_details}{Sao os dados retornados pelo funcao \code{get_procedure_details}}

\item{health_establishment_id}{Código(s) do estabelecimento de saúde}
}
\value{
Um dataset outputSIH
}
\description{
preprocess_SIH: Estrutura os dados de Producao Hospitalar (SIH)
}
