% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_2_preprocess_SIH_RD.R
\name{preprocess_SIH_RD}
\alias{preprocess_SIH_RD}
\title{preprocess_SIH_RD: Estrutura os dados de Producao Hospitalar (SIH)}
\usage{
preprocess_SIH_RD(
  cbo,
  cid,
  raw_SIH_RD,
  county_id,
  procedure_details,
  health_establishment_id
)
}
\arguments{
\item{cbo}{E a tabela retornada pela funcao \code{get_details} com o parametro detail_name igual a "CBO"}

\item{cid}{E a tabela retornada pela funcao \code{get_details} com o parametro detail_name igual a "CID".}

\item{raw_SIH_RD}{Dados do DataSUS de Producao Hospitalar (SIH)}

\item{county_id}{Codigo(s) do Municipio de Atendimento}

\item{procedure_details}{Sao os dados retornados pelo funcao \code{get_procedure_details}}

\item{health_establishment_id}{Código(s) do estabelecimento de saúde}
}
\value{
Um dataset outputSIH_RD
}
\description{
preprocess_SIH_RD: Estrutura os dados de Producao Hospitalar (SIH)
}
