% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_3_preprocess_SIH_RJ.R
\name{preprocess_SIH_RJ}
\alias{preprocess_SIH_RJ}
\title{preprocess_SIH_RJ: Estrutura os dados de Producao Hospitalar (SIH)}
\usage{
preprocess_SIH_RJ(
  cbo,
  cid,
  raw_SIH,
  county_id,
  procedure_details,
  health_establishment_id
)
}
\arguments{
\item{cbo}{E a tabela retornada pela funcao \code{get_details} com o parametro detail_name igual a "CBO"}

\item{cid}{E a tabela retornada pela funcao \code{get_details} com o parametro detail_name igual a "CID".}

\item{raw_SIH}{Dados do DataSUS de Producao Hospitalar (SIH)}

\item{county_id}{Codigo(s) do Municipio de Atendimento}

\item{procedure_details}{Sao os dados retornados pelo funcao \code{get_procedure_details}}

\item{health_establishment_id}{Código(s) do estabelecimento de saúde}
}
\value{
Um dataset outputSIH_RJ
}
\description{
preprocess_SIH_RJ: Estrutura os dados de Producao Hospitalar (SIH)
}
