% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_4_preprocess_SIH_SP.R
\name{preprocess_SIH_SP}
\alias{preprocess_SIH_SP}
\title{Returns data from Professional Services (SP-SIH) in a structured and filtered form.}
\usage{
preprocess_SIH_SP(
  cbo,
  cid,
  raw_SIH_SP,
  county_id,
  procedure_details,
  health_establishment_id
)
}
\arguments{
\item{cbo}{É a tabela retornada pela função \code{get_details} quando o parâmetro \code{detail_name='CBO'}}

\item{cid}{É a tabela retornada pela função \code{get_details} quando o parâmetro \code{detail_name='CID'}}

\item{raw_SIH_SP}{Dados dos Serviços Profissionais (SP) do Sistema de Informação Hospitalar (SIH)}

\item{county_id}{string or a vector of strings. Código do Município de Atendimento.}

\item{procedure_details}{São os dados retornados pelo função \code{get_procedure_details}}

\item{health_establishment_id}{string or a vector of strings. Código do estabelecimento de saúde.}
}
\value{
Retorna a tabela dos Serviços Profissionais (SP)  já filtrada e tratada.
}
\description{
Estrutura e filtra os dados dos Serviços Profissionais (SP) e combina as informações do CNES, SIGTAP e da base de dados \code{counties}.
}
