% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procedure_revoked.R
\name{procedure_revoked}
\alias{procedure_revoked}
\title{Retorna informações sobre procedimentos que foram revogados}
\usage{
procedure_revoked(year_start, month_start, year_end, month_end, dados)
}
\arguments{
\item{year_start}{numeric. Ano inicial para o download dos dados, no formato yyyy.}

\item{month_start}{numeric. Mês inicial para o download dos dados, no formato mm.}

\item{year_end}{numeric. Ano final para o download dos dados, no formato yyyy.}

\item{month_end}{numeric. Mês final para o download dos dados, no formato mm.}

\item{dados}{É o conjunto de dados retornado por uma das funções principais.}
}
\value{
A função retorna um subconjunto da base de dados de entrada, contendo exclusivamente as linhas nas quais a coluna 'Procedimentos realizados' possuía valores nulos. Os valores nulos nestas linhas são imputados com os valores correspondentes.
}
\description{
Alguns procedimentos são revogados, e a partir do mês de revogação, as bases do SIGTAP deixam de fornecer informações sobre eles. Para acessar dados referentes a esses procedimentos, é necessário consultar as informações disponíveis nos meses anteriores à sua revogação.
}
\examples{
  \dontrun{
    dados = procedure_revoked(
      year_start = 2023,
      month_start = 1,
      year_end = 2023,
      month_end = 3,
      dados = outputSIH_RD
    )
  }

}
