% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_3_check_filter.R
\name{check_filter}
\alias{check_filter}
\title{check_filter}
\usage{
check_filter(df, var_value, var_name)
}
\arguments{
\item{df}{um data frame que sera filtrado.}

\item{var_value}{String, Codigo(s) do estabelecimento de saude}

\item{var_name}{String, nome da coluna no data frame que sera usada para a filtragem.}
}
\value{
filtered_df
}
\description{
Funcao auxiliar para filtrar uma lista ou todos; Funcao para filtrar pelo CNES ou pelo municipio
}
