% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7_1_create_output.R
\name{create_output}
\alias{create_output}
\title{Create SUS-SIA/SIH database}
\usage{
create_output(
  year_start,
  month_start,
  year_end,
  month_end,
  state_abbr,
  county_id = NULL,
  health_establishment_id = NULL
)
}
\arguments{
\item{year_start}{Um numero de 4 digitos, indicando o ano de inicio para o download dos dados.}

\item{month_start}{Um numero de 2 digitos, indicando o mes de inicio para o download dos dados.}

\item{year_end}{Um numero de 4 digitos, indicando o ano de termino para o download dos dados.}

\item{month_end}{Um numero de 2 digitos, indicando o mes de termino para o download dos dados.}

\item{state_abbr}{String. Sigla da Unidade Federativa}

\item{county_id}{Codigo(s) do Municipio de Atendimento}

\item{health_establishment_id}{COdigo(s) do estabelecimento de saude}
}
\description{
Processar arquivos dos sistemas de informação SIA e SIH (DATASUS) e combiná-los com informações do CNES e SIGTAP.
}
\examples{
  dados = create_output(
    year_start = 2023,
    month_start = 1,
    year_end = 2023,
    month_end = 3,
    state_abbr = "CE",
    county_id = "230440",
    health_establishment_id = c("2561492", "2481286")
  )
}
