% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_1_create_output_PA.R
\name{create_output_PA}
\alias{create_output_PA}
\title{Create a database for the Ambulatory Information System (SIA/SUS) - PA}
\usage{
create_output_PA(
  year_start,
  month_start,
  year_end,
  month_end,
  state_abbr,
  county_id = NULL,
  health_establishment_id = NULL
)
}
\arguments{
\item{year_start}{numeric. Ano inicial para o download dos dados, no formato yyyy.}

\item{month_start}{numeric. Mês inicial para o download dos dados, no formato mm.}

\item{year_end}{numeric. Ano final para o download dos dados, no formato yyyy.}

\item{month_end}{numeric. Mês final para o download dos dados, no formato mm.}

\item{state_abbr}{string ou vetor de strings. Sigla da Unidade Federativa}

\item{county_id}{string ou vetor de strings. Código do município de atendimento. O padrão é NULL.  Se informado, todos os estabelecimentos de saúde desse município serão filtrados. Este parâmetro é obrigatório se health_establishment_id for NULL.}

\item{health_establishment_id}{string ou vetor de strings. Código do estabelecimento de saúde. O padrão é NULL. Este parâmetro é obrigatório se county_id for NULL. Será desconsiderado se county_id contiver um código válido de município.}
}
\value{
Um DataFrame estruturado contendo dados do SUS-SIA-PA, filtrados por estado ou estabelecimentos de saúde dentro de um intervalo de datas específico, e combinado com informações do CNES e SIGTAP. A função retorna um objeto como os dados e salva a base de dados na pasta './data-raw' em formato CSV, com o nome 'outputSIA.csv'.
}
\description{
Processar arquivos do Sistema de Informação Ambulatorial (SIA) da  Produção Ambulatorial (PA) do DATASUS e integrá-los com dados do CNES e SIGTAP.
}
\examples{
  \dontrun{
    dados = create_output_PA(
      year_start = 2023,
      month_start = 1,
      year_end = 2023,
      month_end = 3,
      state_abbr = "CE",
      county_id = "230440",
      health_establishment_id = c("2561492", "2481286")
    )
  }

}
