% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_2_create_output_SP_from_local.R
\name{create_output_SP_from_local}
\alias{create_output_SP_from_local}
\title{Create SUS-SIH-SP database from local DBC files}
\usage{
create_output_SP_from_local(
  state_abbr,
  dbc_dir_path,
  county_id = NULL,
  health_establishment_id = NULL
)
}
\arguments{
\item{state_abbr}{String. Sigla da Unidade Federativa}

\item{dbc_dir_path}{Diretório que contêm os arquivos DBC}

\item{county_id}{Codigo(s) do Municipio de Atendimento. O padrao é NULL.}

\item{health_establishment_id}{Código(s) do estabelecimento de saúde}
}
\value{
Um DataFrame estruturado contendo dados do SUS-SIH-SP, filtrado por estado ou estabelecimentos de saúde dentro de um intervalo de datas específico, e combinado com informações do CNES e SIGTAP.
}
\description{
Processar arquivos do sistema de informação SIH (DATASUS) que já estão baixados localmente e combiná-los com informações do CNES e SIGTAP.
}
\examples{
  \dontrun{
    create_output_SP_from_local(
      state_abbr = "CE",
      dbc_dir_path = "X:/USID/BOLSA_EXTENSAO_2024/dbc/dbc-2301-2306",
      county_id = "230440",
      health_establishment_id = c("2561492", "2481286")
    )
  }

}
