% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_2_create_output_SP_from_local.R
\name{create_output_SP_from_local}
\alias{create_output_SP_from_local}
\title{Create a database for the Hospital Information System (SIH/SUS) - SP, local data.}
\usage{
create_output_SP_from_local(
  state_abbr,
  dbc_dir_path,
  county_id = NULL,
  health_establishment_id = NULL,
  save_csv = TRUE
)
}
\arguments{
\item{state_abbr}{string ou vetor de strings. Sigla da Unidade Federativa}

\item{dbc_dir_path}{string. Caminho para o diretório  local que contêm os arquivos DBC}

\item{county_id}{string ou vetor de strings. Código do município de atendimento. O padrão é NULL.  Se informado, todos os estabelecimentos de saúde desse município serão filtrados. Este parâmetro é obrigatório se health_establishment_id for NULL.}

\item{health_establishment_id}{string ou vetor de strings. Código do estabelecimento de saúde. O padrão é NULL. Este parâmetro é obrigatório se county_id for NULL. Será desconsiderado se county_id contiver um código válido de município.}

\item{save_csv}{Lógico. O valor padrão é TRUE. Quando definido como TRUE, a base de dados resultante da função é salva como um arquivo CSV no diretório './data-raw'.}
}
\value{
Um DataFrame estruturado contendo dados do SUS-SIH-SP, filtrados por estado ou estabelecimentos de saúde dentro de um intervalo de datas específico, e combinado com informações do CNES e SIGTAP. A função retorna um objeto como os dados e salva a base de dados na pasta './data-raw' em formato CSV, com o nome 'outputSIH_SP.csv'.
}
\description{
Processar arquivos do Sistema de Informação Hospitalar (SIH) dos Serviço Profissionais (SP) do DATASUS que já estão baixados localmente e integrá-los com dados do CNES e SIGTAP.
}
\examples{
    dados = create_output_SP_from_local(
      state_abbr = "CE",
      dbc_dir_path = "X:/USID/BOLSA_EXTENSAO_2024/dbc/dbc-2301-2306",
      county_id = NULL,
      health_establishment_id = c("2561492", "2481286"),
      save_csv = TRUE
    )

}
