% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_1_download_cnes_files.R
\name{download_cnes_files}
\alias{download_cnes_files}
\title{Download data from the National Registry of Health Establishments (CNES)}
\usage{
download_cnes_files(
  year_start,
  month_start,
  year_end,
  month_end,
  newer,
  state_abbr = "all",
  type_data = "ST"
)
}
\arguments{
\item{year_start}{numeric. Ano inicial para o download dos dados, no formato yyyy.}

\item{month_start}{numeric. Mês inicial para o download dos dados, no formato mm.}

\item{year_end}{numeric. Ano final para o download dos dados, no formato yyyy.}

\item{month_end}{numeric. Mês final para o download dos dados, no formato mm.}

\item{newer}{logical. O padrão é TRUE. Se for TRUE e os outros parâmetros forem nulos, obtém o arquivo do mês mais recente disponível no CNES.}

\item{state_abbr}{string. O padrao e "all". Sigla da Unidade Federativa.}

\item{type_data}{string. O padrao e "ST". Valores aceitos ("ST", "LT"). O tipo da base de dados do CNES que sera baixado.}
}
\description{
A função \code{download_cnes_files} baixa arquivos do CNES (Cadastro Nacional de Estabelecimentos de Saúde) e os armazena temporariamente na máquina local.
}
\examples{
\dontrun{download_cnes_files(newer = TRUE)}

}
