% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_download_sigtap.R
\name{download_sigtap_files}
\alias{download_sigtap_files}
\title{download_sigtap_files: Baixa dados .dbc do site DataSUS}
\usage{
download_sigtap_files(
  year_start,
  month_start,
  year_end,
  month_end,
  newer = TRUE,
  specific_dates = NULL
)
}
\arguments{
\item{year_start}{Um numero de 4 digitos, indicando o ano de inicio para o download dos dados.}

\item{month_start}{Um numero de 2 digitos, indicando o mes de inicio para o download dos dados.}

\item{year_end}{Um numero de 4 digitos, indicando o ano de termino para o download dos dados.}

\item{month_end}{Um numero de 2 digitos, indicando o mes de termino para o download dos dados.}

\item{newer}{Logico. O padrao e TRUE. Se for TRUE e os outros parametros forem nulos, obtem o arquivo do mes mais recente disponivel no SIGTAP.}

\item{specific_dates}{Um vetor de strings contendo as datas específicas que se deseja baixar.}
}
\description{
A funcao \code{download_sigtap_files} baixar e descompactar arquivos sigtap do site datasus. Esses arquivos são salvos temporariamente na maquina local. Para cada mes vai ser baixado um arquivo compactado.
}
\examples{
\dontrun{ download_sigtap_files(2024, 1, 2024, 3, newer=FALSE)}

}
