% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_download_sigtap_files.R
\name{download_sigtap_files}
\alias{download_sigtap_files}
\title{Download SIGTAP Files}
\usage{
download_sigtap_files(
  year_start,
  month_start,
  year_end,
  month_end,
  newer = TRUE,
  specific_dates = NULL
)
}
\arguments{
\item{year_start}{numeric. Ano inicial para o download dos dados, no formato yyyy.}

\item{month_start}{numeric. Mês inicial para o download dos dados, no formato mm.}

\item{year_end}{numeric. Ano final para o download dos dados, no formato yyyy.}

\item{month_end}{numeric. Mês final para o download dos dados, no formato mm.}

\item{newer}{logical. O padrão é TRUE. Se for TRUE e os outros parâmetros forem nulos, obtém o arquivo do mês mais recente disponível no SIGTAP.}

\item{specific_dates}{string or a vector of strings. Contem as datas específicas que se deseja baixar.}
}
\description{
A função \code{download_sigtap_files} baixa e descompacta arquivos SIGTAP do site DataSUS. Esses arquivos são salvos temporariamente na máquina local. Para cada mês, será baixado um arquivo compactado. Primeiro, a função verifica se o arquivo do mês especificado já existe na pasta temporária. Se o arquivo estiver presente, ela não baixa o arquivo novamente para esse mês.
}
\examples{
\dontrun{
   download_sigtap_files(2024, 1, 2024, 3, newer = FALSE)
 }
\dontrun{
   download_sigtap_files(specific_dates = c("202210","202301","202303"))}

}
