% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_3_get_CNES.R
\name{get_CNES}
\alias{get_CNES}
\title{get_CNES: Carrega os dados dos estabelecimentos de saude.}
\usage{
get_CNES()
}
\value{
Salva o DataFrame \code{health_establishment} no environment global.
}
\description{
A funcao \code{get_CNES} cria o DataFrame \code{health_establishment}, que contem a uniao de todos os arquivos baixados pela funcao \code{download_cnes_files}.
}
