% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_2_get_counties.R
\name{get_counties}
\alias{get_counties}
\title{get_counties: Obtem informacoes de todos os estados do Brasil}
\usage{
get_counties(state_abbr, county_id)
}
\arguments{
\item{state_abbr}{Sigla da Unidade Federativa}

\item{county_id}{Codigo(s) do Municipio de Atendimento}
}
\value{
Salva o DataFrame \code{counties} no environment global .
}
\description{
A funcao \code{get_counties} cria o DataFrame \code{counties}, que contem informacoes detalhadas sobre os municipios do Brasil.
}
