% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_2_get_counties.R
\name{get_counties}
\alias{get_counties}
\title{Create a dataset with information on all municipalities in Brazil.}
\usage{
get_counties(state_abbr, county_id, download = FALSE)
}
\arguments{
\item{state_abbr}{Sigla da Unidade Federativa}

\item{county_id}{string ou vetor de strings. Código do Município de Atendimento.}

\item{download}{logico. O padrao é FALSE. Se for TRUE, os dados de 'counties' são baixados do site do IBGE. Se for FALSE, os dados são obtidos a partir de uma base de dados incluída no pacote.}
}
\description{
Cria e salva no ambiente global do R o DataFrame \code{counties}, que contém informações detalhadas sobre os municípios do Brasil.
}
