% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_2_get_details.R
\name{get_detail}
\alias{get_detail}
\title{get_detail: Agrupa e estrutura os dados do SIGTAP}
\usage{
get_detail(detail_name)
}
\arguments{
\item{detail_name}{String. Nome do arquivo desejado. Existem 8 opcoes possiveis: "Procedimento", "Grupo", "Subgrupo", "Forma de organizacao", "Financiamento", "Rubrica", "CBO", "CID".}
}
\value{
Uma tabela dos dados correpondentes ao parametro \code{detail_name}
}
\description{
Esta funcao seleciona todos os valores baixados pela funcao \code{download_sigtap_files} que correspondem ao valor passado no parametro \code{detail_name}. Em seguida, transforma esses dados brutos em um formato estruturado e organizado.
}
