% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_1_get_procedure_details.R
\name{get_procedure_details}
\alias{get_procedure_details}
\title{Structure the procedural data downloaded by the download_sigtap_files function into a table.}
\usage{
get_procedure_details()
}
\value{
Retorna a Tabela de Procedimentos do SIGTAP utilizando os dados baixados do SIGTAP.
}
\description{
Esta função agrupa os arquivos de Procedimentos, baixados previamente pela função \code{download_sigtap_files}, consolidando todos os dados em uma única tabela.
}
\examples{
\dontrun{procedure_details = get_procedure_details()}
}
