% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_1_preprocess_SIA.R
\name{preprocess_SIA}
\alias{preprocess_SIA}
\title{preprocess_SIA: Estrutura os dados de Producao Ambulatorial (SIA)}
\usage{
preprocess_SIA(
  cbo,
  cid,
  raw_SIA,
  county_id,
  procedure_details,
  health_establishment_id
)
}
\arguments{
\item{cbo}{E a tabela retornada pela funcao \code{get_details} com o parametro detail_name igual a "CBO"}

\item{cid}{E a tabela retornada pela funcao \code{get_details} com o parametro detail_name igual a "CID".}

\item{raw_SIA}{Dados do DataSUS de Producao Ambulatorial (SIA)}

\item{county_id}{Código(s) do Município de Atendimento}

\item{procedure_details}{Sao os dados retornados pelo funcao \code{get_procedure_details}}

\item{health_establishment_id}{Código(s) do estabelecimento de saúde}
}
\value{
Um dataset outputSIA
}
\description{
preprocess_SIA: Estrutura os dados de Producao Ambulatorial (SIA)
}
