% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_1_preprocess_SIA.R
\name{preprocess_SIA}
\alias{preprocess_SIA}
\title{Returns data from Outpatient Production (PA-SIA) in a structured and filtered form.}
\usage{
preprocess_SIA(
  cbo,
  cid,
  raw_SIA,
  county_id,
  procedure_details,
  health_establishment_id
)
}
\arguments{
\item{cbo}{É a tabela retornada pela função \code{get_details} quando o parâmetro \code{detail_name='CBO'}}

\item{cid}{É a tabela retornada pela função \code{get_details} quando o parâmetro \code{detail_name='CID'}}

\item{raw_SIA}{Dados do DataSUS de Producao Ambulatorial (SIA)}

\item{county_id}{string or a vector of strings. Código do Município de Atendimento.}

\item{procedure_details}{São os dados retornados pelo função \code{get_procedure_details}}

\item{health_establishment_id}{string or a vector of strings. Código do estabelecimento de saúde.}
}
\value{
Retorna a tabela da Produção Ambulatorial já filtrada e tratada.
}
\description{
Estrutura e filtra os dados da Produção Ambulatorial (PA) e combina as informações do CNES, SIGTAP e da base de dados \code{counties}.
}
