% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4_3_preprocess_SIH_RJ.R
\name{preprocess_SIH_RJ}
\alias{preprocess_SIH_RJ}
\title{Returns SIH-AIH Rejected(RJ) data in a structured and filtered form.}
\usage{
preprocess_SIH_RJ(
  cbo,
  cid,
  raw_SIH_RJ,
  county_id,
  procedure_details,
  health_establishment_id
)
}
\arguments{
\item{cbo}{É a tabela retornada pela função \code{get_details} quando o parâmetro \code{detail_name='CBO'}}

\item{cid}{É a tabela retornada pela função \code{get_details} quando o parâmetro \code{detail_name='CID'}}

\item{raw_SIH_RJ}{Dados de Autorização de Internação Hospitalar (AIH) Rejeitadas (RJ) do Sistema de Informação Hospitalar (SIH)}

\item{county_id}{string or a vector of strings. Código do Município de Atendimento.}

\item{procedure_details}{São os dados retornados pelo função \code{get_procedure_details}}

\item{health_establishment_id}{string or a vector of strings. Código do estabelecimento de saúde.}
}
\value{
Retorna a tabela da Autorização de Internação Hospitalar (AIH) Rejeitadas (RJ) já filtrada e tratada.
}
\description{
Estrutura e filtra os dados da Autorização de Internação Hospitalar (AIH) Rejeitadas (RJ) e combina as informações do CNES, SIGTAP e da base de dados \code{counties}.
}
