% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/5_3_preprocess_SIH_SP.R
\name{preprocess_SIH_SP}
\alias{preprocess_SIH_SP}
\title{preprocess_SIH_SP: Estrutura os dados de Producao Hospitalar (SIH-SP)}
\usage{
preprocess_SIH_SP(
  cbo,
  cid,
  raw_SIH_SP,
  county_id,
  procedure_details,
  publication_date_start,
  health_establishment_id
)
}
\arguments{
\item{cbo}{E a tabela retornada pela funcao \code{get_details} com o parametro detail_name igual a "CBO"}

\item{cid}{E a tabela retornada pela funcao \code{get_details} com o parametro detail_name igual a "CID".}

\item{raw_SIH_SP}{Dados do DataSUS de Producao Hospitalar (SIH)}

\item{county_id}{Código(s) do Município de Atendimento}

\item{procedure_details}{Sao os dados retornados pelo funcao \code{get_procedure_details}}

\item{publication_date_start}{E a data de inicio para o download dos dados, formatada no padrao AAAA-MM-DD}

\item{health_establishment_id}{Código(s) do estabelecimento de saúde}
}
\value{
Um dataset outputSIH_SP
}
\description{
preprocess_SIH_SP: Estrutura os dados de Producao Hospitalar (SIH-SP)
}
